link handicap and person, as well as completed Mapper layer for Person. add 'update' for person resource

+++ backend/src/main/java/org/acme/repositories/HandicapPersonRepository.java
@@ -13,6 +13,6 @@ import jakarta.enterprise.context.ApplicationScoped;
 public class HandicapPersonRepository implements PanacheRepositoryBase<HandicapPerson, HandicapPersonId>{
     
     public Optional<HandicapPerson> findByIdOptional(String cin, int handicap_id){
-        return findByIdOptional(new HandicapPersonId(handicap_id, cin));
+        return findByIdOptional(new HandicapPersonId((long)handicap_id, cin));
     }
 }
